+++ client/web/src/Structure/Common/LoadingButton.jsx
@@ -21,7 +21,7 @@ const LoadingButton = ({
       <div className={classes.wrapper}>
         <Button
           variant="contained"
-          style={{ backgroundColor: "#6C8D9E" }}
+          style={{ backgroundColor: "#6C8D9E", color: 'white' }}
           className={buttonClassname}
           disabled={isLoading}
           onClick={submit === "none" ? undefined : submit}
