+++ client/web/src/Structure/Topics/Components/TopicCard.jsx
@@ -40,6 +40,7 @@ const TopicCard = ({ topicData, handleClickOpen, handleOpenProficiency }) => {
           <CardContent
             className={classes.cardStyle}
             onClick={() => handleOpenProficiency({ topic, proficiency })}
+            title="Open Proficiency Data"
           >
             {(lastValue.length && (
               <Typography
@@ -95,6 +96,7 @@ const TopicCard = ({ topicData, handleClickOpen, handleOpenProficiency }) => {
           <CardActions
             className={classes.cardActions}
             onClick={() => handleClickOpen({ topic, topicId })}
+            title={"Attempt Quiz on " + topic}
           >
             <Grid container alignItems="center" spacing={3}>
               <Grid item>
